UI improvments

+++ app/src/main/java/com/example/locations/Admin/all_location/AllLocationsFragment.kt
@@ -1,11 +1,15 @@
 package com.example.locations.Admin.all_location
 
+import android.content.res.ColorStateList
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.activity.OnBackPressedCallback
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
@@ -16,14 +20,19 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.locations.R
 import com.example.locations.Admin.all_location.model.LocationData
 import com.example.locations.databinding.AllLocationsLayoutBinding
+import com.google.ar.sceneform.rendering.Color
+import com.google.firebase.Firebase
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.auth
 import java.util.Locale
 
 class AllLocationsFragment : Fragment() {
 
     // Binding object instance corresponding to the fragment_all_locations.xml layout
     private var _binding: AllLocationsLayoutBinding? = null
-
-    // Non-null access to the binding object instance
+    private lateinit var auth: FirebaseAuth
+    private var admin="navigationproject2024@gmail.com"
+   // Non-null access to the binding object instance
     private val binding get() = _binding!!
 
     // Adapter for the RecyclerView
@@ -38,13 +47,34 @@ class AllLocationsFragment : Fragment() {
                               savedInstanceState: Bundle?):
             View {
         _binding = AllLocationsLayoutBinding.inflate(inflater, container, false)
+        binding.btnBack.setOnClickListener {
+            returnToAdd()
+        }
+        val colorStateList = ColorStateList.valueOf(ContextCompat.getColor(requireContext(), R.color.gray))
+        binding.searchEditTextLayout.defaultHintTextColor = colorStateList
         // Setup RecyclerView
-        setupRecyclerView()
+        auth= Firebase.auth
+        val currentUser=auth.currentUser?.email.toString()
+       setupRecyclerView(currentUser)
+
         // Setup back button click listener
 
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner,object : OnBackPressedCallback(true) {
+            override fun handleOnBackPressed() {
+               returnToAdd()
+            }
+
+        })
         return binding.root
     }
-
+private fun returnToAdd(){
+    val returnToFragmentId=arguments?.getInt("returnToFragmentId")
+    if(returnToFragmentId!=null) {
+    findNavController().navigate(returnToFragmentId)
+    }else{
+        findNavController().navigate(R.id.action_allLocationsFragments_to_StartNav)
+    }
+}
     // Called immediately after onCreateView() has returned, but before any saved state has been restored in to the view
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -53,10 +83,10 @@ class AllLocationsFragment : Fragment() {
     }
 
     // Setup RecyclerView with data from the ViewModel
-    private fun setupRecyclerView() {
+    private fun setupRecyclerView(currentUser:String) {
         viewModel.locationData?.observe(viewLifecycleOwner) { allLocations ->
             setupAdapter(allLocations)
-            setupItemTouchHelper()
+            setupItemTouchHelper(currentUser)
         }
     }
 
@@ -71,7 +101,7 @@ class AllLocationsFragment : Fragment() {
             override fun onItemLongClicked(index: Int) {
                 viewModel.setLocation(allLocations[index])
                 setupAdapter(allLocations)
-                binding.searchEditText.text.clear()
+                binding.searchEditTextLayout.editText?.text?.clear()
                 findNavController().navigate(R.id.action_allLocationsFragments_to_detailLocationInfo)
             }
         })
@@ -81,17 +111,28 @@ class AllLocationsFragment : Fragment() {
     }
 
     // Setup item touch helper for RecyclerView
-    private fun setupItemTouchHelper() {
+    private fun setupItemTouchHelper(currentUser:String) {
         ItemTouchHelper(object : ItemTouchHelper.Callback() {
             // Setup movement flags for RecyclerView
-            override fun getMovementFlags(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder) =
-                makeFlag(ItemTouchHelper.ACTION_STATE_SWIPE, ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT)
+            override fun getMovementFlags(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder): Int{
+                if(currentUser==admin) {
+              return makeFlag(
+                    ItemTouchHelper.ACTION_STATE_SWIPE,
+                    ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT
+                )
+                }else{
+                    return 0
+                }
+            }
             // Handle move event
             override fun onMove(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder, target: RecyclerView.ViewHolder) = false
             // Handle swipe event
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                val location = (binding.recyclerView.adapter as LocationAdapter).itemAt(viewHolder.adapterPosition)
-                viewModel.deleteEntry(location)
+                if(currentUser==admin) {
+                    val location =
+                        (binding.recyclerView.adapter as LocationAdapter).itemAt(viewHolder.adapterPosition)
+                    viewModel.deleteEntry(location)
+                }
             }
         }).attachToRecyclerView(binding.recyclerView)
     }
